refactor: seperated unit and end-to-end test
